file.py: Rename to local_target.py
 * Also remove deprecated File class
