+++ luigi/__init__.py
@@ -24,8 +24,8 @@ from luigi.task import Task, Config, ExternalTask, WrapperTask, namespace
 from luigi import target
 from luigi.target import Target
 
-from luigi import file  # wtf @ naming
-from luigi.file import File, LocalTarget
+from luigi import local_target
+from luigi.local_target import LocalTarget
 
 from luigi import rpc
 from luigi.rpc import RemoteScheduler, RPCError
@@ -52,12 +52,12 @@ from .tools import range  # just makes the tool classes available from command l
 
 __all__ = [
     'task', 'Task', 'Config', 'ExternalTask', 'WrapperTask', 'namespace',
-    'target', 'Target', 'File', 'LocalTarget', 'rpc', 'RemoteScheduler',
+    'target', 'Target', 'LocalTarget', 'rpc', 'RemoteScheduler',
     'RPCError', 'parameter', 'Parameter', 'DateParameter', 'MonthParameter',
     'YearParameter', 'DateHourParameter', 'DateMinuteParameter', 'DateSecondParameter', 'range',
     'DateIntervalParameter', 'TimeDeltaParameter', 'IntParameter',
     'FloatParameter', 'BooleanParameter', 'BoolParameter', 'TaskParameter',
     'ListParameter', 'TupleParameter', 'EnumParameter', 'DictParameter',
-    'configuration', 'interface', 'file', 'run', 'build', 'event', 'Event',
+    'configuration', 'interface', 'local_target', 'run', 'build', 'event', 'Event',
     'NumericalParameter', 'ChoiceParameter'
 ]
